requirements.txt
appnope==0.1.0
attrs==16.0.0
backports.shutil-get-terminal-size==1.0.0
cffi==1.7.0
cryptography==1.4
cssselect==0.9.2
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.6
django-annoying==0.9.0
django-bootstrap3==7.0.1
django-ckeditor==5.1.1
django-cors-headers==1.1.0
django-dotenv==1.4.1
django-dynamic-scraper==0.11.0
django-eventtools==0.9.7
django-filter==0.13.0
django-localflavor==1.3
django-recurrence==1.3.0
django-scheduler==0.8
django-timezone-field==2.0
djangorestframework==3.3.3
enum34==1.1.6
future==0.15.2
gdal==1.11.1
gnureadline==6.3.3
googlemaps==2.4.3
gunicorn==19.6.0
icalendar==3.9.2
idna==2.1
ipaddress==1.0.16
ipdb==0.10.1
ipython==4.2.0
ipython-genutils==0.1.0
jsonpath-rw==1.4.0
lxml==3.6.0
parsel==1.0.2
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
ply==3.8
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.0.0
python-dateutil==2.5.3
pytz==2016.4
queuelib==1.4.2
redis==2.10.5
requests==2.9.1
Scrapy==1.1.0
scrapy-djangoitem==1.1.1
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.1
traitlets==4.2.1
Twisted==16.2.0
w3lib==1.14.2
wheel==0.24.0
whitenoise==3.2
zope.interface==4.2.0
